component ContentPlaceholder add failing test
